+++ lekin/lekin_struct/resource.py
@@ -4,6 +4,7 @@ Resource/Machine Struct
 
 import math
 
+import numpy as np
 import pandas as pd
 
 from lekin.lekin_struct.timeslot import TimeSlot
@@ -67,7 +68,7 @@ class Resource:
 
     def get_latest_available_time(self, duration=None, end=None):
         self.update_continuous_empty_hours()
-        return max([i for i in self.continuous_empty_hours[:end] if i >= duration])
+        return max([i + 1 for (i, v) in enumerate(self.continuous_empty_hours[:end]) if v >= duration])
 
     def update_continuous_empty_hours(self):
         if len(self.available_hours) != len(self._available_timeslots):