+++ lekin/scheduler.py
@@ -10,12 +10,12 @@ from lekin.datasets.check_data import check_data
 
 
 class Scheduler(object):
-    def __init__(self, objective, solver, max_operations):
+    def __init__(self, objective, solver, max_operations, **kwargs):
         self.objective = objective
         self.solver = solver
         self.max_operations = max_operations
 
-    def solve(self, jobs, machines):
+    def run(self, jobs, machines):
         self.solver.solve(jobs, machines)
 
     def evaluate(self):