+++ lekin/solver/construction_heuristics/backward.py
@@ -3,12 +3,18 @@
 import logging
 import math
 
-from lekin.lekin_struct.timeslot import TimeSlot
+from lekin.lekin_struct import JobCollector, ResourceCollector, RouteCollector, TimeSlot
 from lekin.solver.construction_heuristics.base import BaseScheduler
 
 
 class BackwardScheduler(object):
-    def __init__(self, job_collector, resource_collector, route_collector=None, **kwargs):
+    def __init__(
+        self,
+        job_collector: JobCollector,
+        resource_collector: ResourceCollector,
+        route_collector: RouteCollector = None,
+        **kwargs,
+    ):
         self.job_collector = job_collector
         self.resource_collector = resource_collector
         self.route_collector = route_collector
@@ -39,7 +45,7 @@ class BackwardScheduler(object):
 
         op_earliest_start = 0  # forward constraint
         op_latest_end = 150  # backward constraint
-        for operation in job.operations[::-1]:  # inverse
+        for operation in job.operations[::-1]:  # inverse for backward
             logging.info(f"\tAssign Operation {operation.operation_id} of Job {job.job_id}")
             chosen_resource, chosen_timeslot_hour = self.find_best_resource_and_timeslot_for_operation(
                 operation, op_latest_end, op_earliest_start
@@ -76,9 +82,10 @@ class BackwardScheduler(object):
                 latest_index = i
                 resource_latest_time = resource_time
 
+        # print(operation.operation_id, operation.processing_time, op_latest_end, resource_latest_time)
         chosen_resource = available_resource[latest_index]
         latest_time = int(min(op_latest_end, resource_latest_time))
-        chosen_hours = list(range(latest_time - math.ceil(operation.processing_time), latest_time + 0))
+        chosen_hours = list(range(latest_time - math.ceil(operation.processing_time), latest_time + 1))
         return chosen_resource, chosen_hours
 
     def assign_operation(self, operation, start_time, end_time, resources):